[wip] Prefer to use the libiconv library.
Previously, it didn't search for the libiconv unless it had an additional libiconv library it's self build.
On Mac OS environment, Use the libiconv rather than the Core Foundation.

If you does not want to use the libiconv (instead to use the Core Foundation explicityly on Mac),
you have to do the followings

  ./configure --disable-iconv