Change behavior of overwriting symbolic link during extraction.
Current behavior of overwriting a symbolic link with a regular file
depends on a file type the symbolic link refers to.
When it refers to a regular file, lha asks whether to overwrite or not.
When it refers to a directory or a special file, lha fails to overwrite.
When it refers to broken link, lha silently overwrite.

The fix changes to ask whether to overwrite for all these cases.